+++ test/bigtest/tests/piece/unreceive-piece-test.js
@@ -16,8 +16,16 @@ describe('Unreceive piece', () => {
   setupApplication();
 
   beforeEach(async function () {
+    const vendor = this.server.create('vendor');
+    const order = this.server.create('order', {
+      vendor: vendor.id,
+    });
     const line = this.server.create('line', {
       orderFormat: ORDER_FORMATS.physicalResource,
+      purchaseOrderId: order.id,
+      physical: {
+        materialSupplier: vendor.id,
+      },
     });
     const title = this.server.create('title', {
       poLineId: line.id,