+++ test/bigtest/tests/title/title-details-test.js
@@ -20,17 +20,22 @@ describe('Title details', () => {
   setupApplication();
 
   beforeEach(async function () {
+    const vendor = this.server.create('vendor');
     const order = this.server.create('order', {
       workflowStatus: ORDER_STATUSES.closed,
       closeReason: {
         reason: 'Reason for closure',
       },
+      vendor: vendor.id,
     });
     const line = this.server.create('line', {
       orderFormat: ORDER_FORMATS.physicalResource,
       purchaseOrderId: order.id,
       details: {
         receivingNote: 'Receiving note',
+      },
+      physical: {
+        materialSupplier: vendor.id,
       },
     });
     const title = this.server.create('title', {